+++ apps/engine/src/modules/pnl.rs
@@ -1,9 +1,9 @@
 use crate::modules::types::{Order, Side};
 
 /// Calculate the profit or loss (PnL) for an order
-pub fn calculate_pnl(order: &Order, close_price: &i64) -> i64 {
-    let entry_price = order.price.unwrap();
-    let quantity = order.quantity;
+pub fn calculate_pnl(order: &Order, close_price: &f64) -> f64 {
+    let entry_price = order.price.unwrap() as f64;
+    let quantity = order.quantity as f64;
 
     match order.side {
         Side::Buy => (close_price - entry_price) * quantity / entry_price,
